pyproject.toml
[tool.poetry]
name = "scp-bridge"
version = "0.1.0"
description = ""
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
watchdog = "^4.0.2"
google-cloud-storage = "^2.18.2"
loguru = "^0.7.2"
pydantic = "^2.8.2"
pendulum = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
